refactoring some code and test

+++ frontend/src/components/List/commons.js
@@ -13,17 +13,24 @@ const useContactList = () => {
   }, []);
 
   const handleDelete = (id) => {
-    contactsApi.deleteContact({ id: id }).then(() => {
-      var newArr = contacts.filter(function (obj) {
-        return obj.id !== id;
+    if (window.confirm("Are you sure you wish to delete this item?"))
+      contactsApi.deleteContact({ id: id }).then(() => {
+        var newArr = contacts.filter(function (obj) {
+          return obj.id !== id;
+        });
+        setContacts(newArr);
       });
-      setContacts(newArr);
-    });
+  };
+
+  const handleUpdate = (e, contact) => {
+    e.preventDefault();
+    window.location.href = "/contact/" + contact.id;
   };
 
   return {
     contacts,
     handleDelete,
+    handleUpdate,
   };
 };
 export default useContactList;
